==> analysis/part2.Rmd <==
---
title: "part2"
output:
  workflowr::wflow_html:
    code_folding: hide
    toc: true
---

```{r, echo=F, message=F, warning=F}
rm(list=ls())
set.seed(123)
knitr::opts_chunk$set(autodep = TRUE)
library(knitr)
library(RSQLite)
library(dplyr)
library(knitr)
library(tidyverse)
library(gridExtra)
library(ggpubr)
```


```{r, echo=F, message=F, warning=F, results=F, eval=T}
multiplesheets <- function(fname) {
   
  # getting info about all excel sheets
  sheets <- readxl::excel_sheets(fname)
  tibble <- lapply(sheets, function(x) readxl::read_excel(fname, sheet = x))
  data_frame <- lapply(tibble, as.data.frame)
    
  # assigning names to data frames
  names(data_frame) <- sheets
    
  # print data frame
  print(data_frame)
}

####
# load SCZ gene 
SCHEMA_gene_results=multiplesheets("C:\\han\\Dataset\\SCHEMA\\41586_2022_4556_MOESM3_ESM.xlsx")
scz1=data.frame(Gene=SCHEMA_gene_results$`Table S5 - Gene Results`$`Gene Symbol`, pval.meta=SCHEMA_gene_results$`Table S5 - Gene Results`$`P meta`, qval.meta=SCHEMA_gene_results$`Table S5 - Gene Results`$`Q meta`, dn_PTV=SCHEMA_gene_results$`Table S5 - Gene Results`$`De novo PTV`) 
scz1$dn_PTV[is.na(scz1$dn_PTV)]=0   # fill missing PTV with 0 
scz1=scz1%>% drop_na()  # remove NA's  ######## this is important 

#scz2=scz1[-c(9934,  9935,  9936, 13100),]   #  remove duplicates which(duplicated(scz1$Gene)==T)
#unique_gene= unique(scz1$Gene)
#scz=data.frame(Gene=unique_gene, pval.meta=scz1$pval.meta[which(scz1$Gene %in% unique_gene)])  # remove duplicates 
```




```{r, message=F, warning=F}
load("C:/han/Projects/Multi_Omics/multi_omics/data/gene_pfc35_20241203.RData")

SCHEMA_top_10_gene=scz1$Gene[1:10]
SCHEMA_FDR_gene=scz1$Gene[1:32]

#X=list(
#SCHEMA_10_Risk_Gene=SCHEMA_top_10_gene,
#DAWN_gene=genes_pfc35,
#SCHEMA_FDR_gene=SCHEMA_FDR_gene
#  )


#library(ggvenn)
#ggvenn(
#  X, 
#  fill_color = c("#0073C2FF", "#EFC000FF",  "#CD534CFF"),
#  stroke_size = 0.5, set_name_size = 4
#  )

#png("C:\\Users\\hans\\OneDrive - Marquette University\\SCZ-DAWN\\Figures\\Fig_venn_diagram.png")

library("VennDiagram") 
# move to new plotting page 
grid.newpage() 
  
# create Venn diagram with three sets 
draw.triple.venn(area1=10,area2=47, area3 =32,  
                 n12=9, n23=28, n13=10, n123=9,  
                 category=c("SCHEMA 10 risk genes","DAWN genes","SCHEMA FDR genes"), 
                 col="red",fill=c("Green","Orange","Blue"), lty="solid", 
                 cat.cex = 1.3, cat.pos=c(0,0,180), cat.col=c("Green","Orange","Blue"), cex=1.5, scaled = F) # cat.cex  the size of the category names; cex: A vector (length 7) giving the size of the areas' labels; cat.pos: position of category names https://www.rdocumentation.org/packages/VennDiagram/versions/1.7.3/topics/draw.triple.venn
#dev.off()
#write.table(genes_pfc35, file="C:\\Users\\hans\\OneDrive - Marquette University\\SCZ-DAWN\\genes_pfc35.txt",row.names=F, col.names = F, quote = F) # output this gene list as input for GO enrichment analysis 

#write.table(setdiff(genes_pfc35,SCHEMA_FDR_gene), file="C:\\Users\\hans\\OneDrive - Marquette University\\SCZ-DAWN\\new_genes_pfc35.txt",row.names=F, col.names = F, quote = F) # output this gene list as input for GO enrichment analysis

#write.table(intersect(genes_pfc35,SCHEMA_FDR_gene), file="C:\\Users\\hans\\OneDrive - Marquette University\\SCZ-DAWN\\28SCHEMA_FDR_gene.txt",row.names=F, col.names = F, quote = F) # output this gene list as input for GO enrichment analysis
```

* a better way to draw venn diagram might be using online tool [BioVenn](https://www.biovenn.nl/index.php)

* 4 SCHEMA FDR genes ("SETD1A"  "FAM178A" "OR4P4"   "PCLO" ) are missed by DAWN is because these 4 genes don't have expression data in brain span. 

* 19 new genes are found 

* enrichment with 10 SCHEMA risk genes, significantly enriched because of p value of 3.75e-12. 

```{r, message=F, warning=F}
fisher.test(matrix(c(3100, 47, 9, 9), nrow=2))
```


* enrichment with 32 SCHEMA FDR genes, significantly enriched because of p value less than 2.2e-16.

```{r, message=F, warning=F}
fisher.test(matrix(c(3100, 47, 28, 28), nrow=2))
```
