==> analysis/part3.Rmd <==
---
title: "enrichment with other gene sets"
output:
  workflowr::wflow_html:
    code_folding: hide
    toc: true
---


```{r, echo=F, message=F, warning=F}
rm(list=ls())
set.seed(123)
knitr::opts_chunk$set(autodep = TRUE)
library(knitr)
library(RSQLite)
library(dplyr)
library(knitr)
library(tidyverse)
library(gridExtra)
library(ggpubr)
```



```{r, echo=F, message=F, warning=F, results=F, eval=T}
multiplesheets <- function(fname) {
   
  # getting info about all excel sheets
  sheets <- readxl::excel_sheets(fname)
  tibble <- lapply(sheets, function(x) readxl::read_excel(fname, sheet = x))
  data_frame <- lapply(tibble, as.data.frame)
    
  # assigning names to data frames
  names(data_frame) <- sheets
    
  # print data frame
  print(data_frame)
}

####
# load SCZ gene 
SCHEMA_gene_results=multiplesheets("C:\\han\\Dataset\\SCHEMA\\41586_2022_4556_MOESM3_ESM.xlsx")
scz1=data.frame(Gene=SCHEMA_gene_results$`Table S5 - Gene Results`$`Gene Symbol`, pval.meta=SCHEMA_gene_results$`Table S5 - Gene Results`$`P meta`, qval.meta=SCHEMA_gene_results$`Table S5 - Gene Results`$`Q meta`, dn_PTV=SCHEMA_gene_results$`Table S5 - Gene Results`$`De novo PTV`) 
scz1$dn_PTV[is.na(scz1$dn_PTV)]=0   # fill missing PTV with 0 
scz1=scz1%>% drop_na()  # remove NA's  ######## this is important 

scz2=scz1[-c(9934,  9935,  9936, 13100),]   #  remove duplicates which(duplicated(scz1$Gene)==T)
#unique_gene= unique(scz1$Gene)
#scz=data.frame(Gene=unique_gene, pval.meta=scz1$pval.meta[which(scz1$Gene %in% unique_gene)])  # remove duplicates 
```


## Venn diagrams of genes shared across different disorders 


```{r, message=F, warning=F}
load("C:/han/Projects/Multi_Omics/multi_omics/data/gene_pfc35_20241203.RData") # load gene_pfc35
load("C:/han/Projects/Multi_Omics/multi_omics/data/dat_pfc35_20241203.RData") # load data dat_pfc35
Fu_gene_list=as_tibble(read.csv("C:/han/Dataset/20231221_PMID_35982160/Fu 2022 Suppl-41588_2022_1104_MOESM3_ESM.csv"))
ASC_gene=Fu_gene_list %>% filter(ASD185=="TRUE") %>% dplyr::select(gene_gencodeV33) %>% pull
neuro_disorers=SCHEMA_gene_results$`Table S11 - Genes from genome-w`
DD_ID_gene=as_tibble(neuro_disorers) %>% filter(Category=="DD/ID genes") %>% dplyr::select(`Gene name`) %>% pull()
SCZ_GWAS_gene=as_tibble(neuro_disorers) %>% filter(Category=="Schizophrenia GWAS genes") %>% dplyr::select(`Gene name`) %>% pull()

x=list(
  DAWN_gene=genes_pfc35,
  #Autism_gene=as_tibble(neuro_disorers) %>% filter(Category=="ASD genes") %>% dplyr::select(`Gene name`) %>% pull(),
  Autism_gene=ASC_gene,
  DD_ID_gene=DD_ID_gene,
  SCZ_GWAS_gene=SCZ_GWAS_gene

)
library(VennDiagram)

display_venn <- function(x, ...){
  library(VennDiagram)
  grid.newpage()
  venn_object <- venn.diagram(x, filename = NULL, ...)
  grid.draw(venn_object)
}

#png("C:\\Users\\hans\\OneDrive - Marquette University\\SCZ-DAWN\\Figures\\Fig_overlapping_genes.png")

display_venn(
        x,
    #    category.names = c("MIRAGE FDR genes" , "Autism genes" , "DD/ID genes", "SCZ GWAS genes"),
    category.names = c("DAWN genes" , "Autism genes" , "DD/ID genes", "SCZ GWAS genes"),
        # Circles
        lwd = 2,
        lty = 'blank',
     #   fill = c("#999999", "#E69F00", "#56B4E9", "#009E73"),
    fill = c("red", "blue", "green", "yellow"),
        # Numbers
        cex = 1.5,
        fontface = "italic",
        # Set names
        cat.cex = 1.5,
        cat.fontface = "bold",
        cat.default.pos = "outer",
       # cat.dist = c(0.055, 0.055, 0.1, 0.1, 0.3 )
         cat.dist = c(0.055, 0.055, 0.1, 0.1 )
)

#dev.off()
```




### DAWN genes overlpping with autism 


#### overlapping genes between DAWN 46 genes and ASC genes  


```{r,message=F, warning=F}
intersect(genes_pfc35, ASC_gene)
fisher.test(matrix(c(3100, 47, length(intersect(colnames(dat_pfc35), ASC_gene)), length(intersect(genes_pfc35, ASC_gene))), nrow=2))
```
#### overlapping genes between SCHEMA FDR genes and ASC genes 


```{r, message=F, warning=F}
SCHEMA_FDR_gene=SCHEMA_gene_results$`Table S5 - Gene Results`[order(SCHEMA_gene_results$`Table S5 - Gene Results`$`P ca/co (comb)`, decreasing = F),]$`Gene Symbol`[1:32]
intersect(SCHEMA_FDR_gene, ASC_gene)  # overlapping genes between SCHEMA and 185 autism genes 
```

#### new overlapping genes between DAWN 46 genes and ASC genes 

```{r, message=F, warning=F}
setdiff(intersect(genes_pfc35, ASC_gene),intersect(SCHEMA_FDR_gene, ASC_gene)) 
```



### DAWN genes overlpping with DD/ID 

```{r, message=F, warning=F}
intersect(genes_pfc35, DD_ID_gene)
fisher.test(matrix(c(3100, 47, length(intersect(colnames(dat_pfc35), DD_ID_gene)), length(intersect(genes_pfc35, DD_ID_gene))), nrow=2))
intersect(intersect(genes_pfc35, DD_ID_gene), ASC_gene)
intersect(intersect(genes_pfc35, DD_ID_gene), SCZ_GWAS_gene)
```

* 10 DD/ID genes shared with SCHEMA FDR genes: SETD1A, TRIO, CACNA1G, GRIN2A, SRRM2, ZMYM2, KDM6B, ASH1L, STAG1, HIST1H1E, from SCHEMA Supp table 

```{r, message=F, warning=F}
DD_ID_SCHEMA=c("SETD1A", "TRIO", "CACNA1G", "GRIN2A", "SRRM2", "ZMYM2", "KDM6B", "ASH1L", "STAG1", "HIST1H1E")
setdiff(intersect(genes_pfc35, DD_ID_gene), DD_ID_SCHEMA)
```

* two new genes NR4A2, ZMYND11 overlapping with DD/ID, compared to SCHEMA results. 



### DAWN genes overlpping with SCZ GWAS gene 

```{r, echo=T, message=F, warning=F}
intersect(genes_pfc35, SCZ_GWAS_gene)
fisher.test(matrix(c(3100, 47, length(intersect(colnames(dat_pfc35), SCZ_GWAS_gene)), length(intersect(genes_pfc35, SCZ_GWAS_gene))), nrow=2))
```


## OR of DAWN genes in different gene sets 


```{r, message=F, warning=F}

# enrichment analysis
all_gene=colnames(dat_pfc35)
Autism_gene=ASC_gene
DD_ID_gene=as_tibble(neuro_disorers) %>% filter(Category=="DD/ID genes") %>% dplyr::select(`Gene name`) %>% pull()
SCZ_GWAS_gene=as_tibble(neuro_disorers) %>% filter(Category=="Schizophrenia GWAS genes") %>% dplyr::select(`Gene name`) %>% pull()
SCHEMA_FDR_gene=SCHEMA_gene_results$`Table S5 - Gene Results`[order(SCHEMA_gene_results$`Table S5 - Gene Results`$`P ca/co (comb)`, decreasing = F),]$`Gene Symbol`[1:32]
dawn_new_gene=setdiff(genes_pfc35, SCHEMA_FDR_gene)

gene_set=list(all_gene=all_gene, Autism_gene=Autism_gene, DD_ID_gene=DD_ID_gene, SCZ_GWAS_gene=SCZ_GWAS_gene, SCHEMA_FDR_gene=SCHEMA_FDR_gene, DAWN_gene=genes_pfc35, DAWN_new_gene=dawn_new_gene)

##########################
OR=numeric(); pvalue_dawn=numeric(); pvalue_dawn_new=numeric(); pvalue_schema=numeric()
percentage=matrix(nrow=4, ncol=3)
rownames(percentage)=c("dawn_gene", "all_gene", "schema_fdr", "dawn_new_gene")
colnames(percentage)=names(gene_set)[2:4]
DAWN_overlap_matrix=list()
DAWN_new_overlap_matrix=list()
for (i in 2:4)
{
dawn_matrix=matrix(c(length(intersect(gene_set$DAWN_gene, gene_set[[i]])), 
                                 length(intersect(gene_set$all_gene, gene_set[[i]])), 
                                 length(gene_set$DAWN_gene), 
                                 length(gene_set$all_gene)), nrow=2)
  fisher_test=fisher.test(dawn_matrix)
  OR[i-1]=fisher_test$estimate; pvalue_dawn[i-1]=fisher_test$p.value; 
  DAWN_overlap_matrix[[i-1]]=dawn_matrix
  
  dawn_new_matrix=matrix(c(length(intersect(gene_set$DAWN_new_gene, gene_set[[i]])), 
                                        length(intersect(gene_set$all_gene, gene_set[[i]])), 
                                        length(gene_set$DAWN_new_gene), 
                                        length(gene_set$all_gene)), nrow=2)
pvalue_dawn_new[i-1]=fisher.test(dawn_new_matrix)$p.value
DAWN_new_overlap_matrix[[i-1]]=dawn_new_matrix
  
 percentage[1,(i-1)]=length(intersect(gene_set$DAWN_gene, gene_set[[i]]))/length(gene_set$DAWN_gene)
 percentage[2,(i-1)]=length(intersect(gene_set$all_gene, gene_set[[i]]))/length(gene_set$all_gene)
 percentage[3,(i-1)]=length(intersect(gene_set$SCHEMA_FDR_gene, gene_set[[i]]))/length(gene_set$SCHEMA_FDR_gene)
 percentage[4,(i-1)]=length(intersect(gene_set$DAWN_new_gene, gene_set[[i]]))/length(gene_set$DAWN_new_gene)
                                                                                                                   
 pvalue_schema[i-1]=fisher.test(matrix(c(length(intersect(gene_set$SCHEMA_FDR_gene, gene_set[[i]])), 
                                 length(intersect(gene_set$all_gene, gene_set[[i]])), 
                                 length(gene_set$SCHEMA_FDR_gene), 
                                 length(gene_set$all_gene)), nrow=2))$p.value
 
}

```


```{r, echo=F, message=F, warning=F}
OR_pvalue=data.frame(OR=OR, pvalue=round(pvalue_dawn,4), genes=c( "Autism genes", "DD/ID genes", "SCZ GWAS genes"))

ggplot(data=OR_pvalue, aes(x=genes, y=OR, fill=genes)) + 
     geom_bar(position = 'dodge', stat='identity') +
     xlab("")+ylab("odds ratio")+
  theme(plot.title = element_text(hjust = 0.5, size=15, face="bold"))+
  ggtitle("OR of DAWN genes in different gene sets")+
  theme(legend.position="none")+
     geom_text(aes(label= pvalue), position=position_dodge(width=0.9), vjust=-0.25)
```

* numbers over bars are p values. 



## enrichment of DAWN (new) genes with other gene sets  


```{r, message=F, warning=F}

percentage_plot=data.frame(prop=c(percentage[2,],percentage[1,], percentage[3,]), genes=rep(c("Autism genes", "DD/ID genes",    "SCZ GWAS genes"),3), group=rep(c("all genes", "DAWN genes", "SCHEMA FDR genes"), each=3))

textsize=5
#                                                                                                 par(mar=rep(0,4))
topmirage_enrichment=
  ggplot(percentage_plot, aes(genes, prop)) + 
   geom_bar(aes(fill = group), stat = "identity", position = "dodge")+
   theme_classic()+
   xlab("")+ylab("Proportion of overlapping genes in the gene set")+
  theme(axis.text.y = element_text(size = 15))+ # this is for numbers on y axis 
   ylim(c(0, 0.35))+
   theme(plot.title = element_text(hjust = 0.5, size=15, face="bold"))+
   ggtitle("Enrichment of DAWN genes with other gene sets")+
    theme(legend.title = element_blank())+
   geom_signif(annotations ="0.005", y_position = 0.14  ,xmin=0.7, xmax=1, size=1, vjust=-0.5, textsize=textsize)+ # textsize is for text over the bar; size is for width of lines 
  geom_signif(annotations ="0.020", y_position = 0.18  ,xmin=0.7, xmax=1.3, size=1, vjust=-0.5, textsize=textsize)+ # textsize is for text over the bar; size is for width of lines
     geom_signif(annotations ="6e-06", y_position = 0.325  ,xmin=1.7, xmax=2.3, size=1, vjust=-0.5, textsize=textsize)+
  geom_signif(annotations ="1e-04", y_position = 0.23  ,xmin=1.7, xmax=2, size=1, vjust=-0.5, textsize=textsize)+
   geom_signif(annotations ="0.002", y_position = 0.09  ,xmin=2.7, xmax=3.3, size=1, vjust=-0.5, textsize=textsize)+
  geom_signif(annotations ="0.034", y_position = 0.052  ,xmin=2.7, xmax=3.0, size=1, vjust=-0.5, textsize=textsize)+
   theme(legend.position="bottom")+
   theme(axis.text.x = element_text(angle = 0, vjust = 0.5, size=15))+
 theme(text = element_text(size=20))+ # size is for labels on x axis and legend title 
  theme(axis.title = element_text(size = 13))+ # Font Size of Axis Titles
 theme(legend.spacing.x = unit(1.0, 'cm')) # add space between symbols in legend 
###################### compute p value 

#png("C:\\Users\\hans\\OneDrive - Marquette University\\SCZ-DAWN\\Figures\\Fig-enrichment_of_top_genes.png")
topmirage_enrichment
#dev.off()
```


  
```{r, echo=F, message=F, warning=F, eval=F}
percentage=data.frame(prop=c(all_gene_percentage, dawn_percentage), gene_set=c("DD/ID genes", "SCZ GWAS genes", "Autism genes", "DD/ID genes", "SCZ GWAS genes", "Autism genes"), group=rep(c("all genes", "DAWN genes"), each=3))


topmirage_enrichment=
  ggplot(percentage, aes(gene_set, prop)) + 
   geom_bar(aes(fill = group), stat = "identity", position = "dodge")+
   theme_classic()+
   xlab("")+ylab("Proportion of genes in the gene set")+
  theme(axis.text.y = element_text(size = 13))+ # this is for numbers on y axis 
   ylim(c(0, 0.25))+
   theme(plot.title = element_text(hjust = 0.5, size=15, face="bold"))+
   ggtitle("Enrichment of DAWN genes with other gene sets")+
    theme(legend.title = element_blank())+
   geom_signif(annotations ="0.007", y_position = 0.125  ,xmin=0.7, xmax=1.3, size=1, vjust=-0.5, textsize=5.0)+ # textsize is for text over the bar; size is for width of lines 
     geom_signif(annotations ="0.003", y_position = 0.16  ,xmin=1.7, xmax=2.3, size=1, vjust=-0.5, textsize=5.0)+
   geom_signif(annotations ="0.083", y_position = 0.02  ,xmin=2.7, xmax=3.3, size=1, vjust=-0.5, textsize=5.0)+
   theme(legend.position="bottom")+
  theme(legend.text=element_text(size=15))+
   theme(axis.text.x = element_text(angle = 0, vjust = 0.5, size=15))+
 theme(text = element_text(size=15))+ # size is for labels on x axis 
  theme(axis.title = element_text(size = 15))+ # Font Size of Axis Titles
 theme(legend.spacing.x = unit(1.0, 'cm')) # add space between symbols in legend

#png("C:/Users/hans/OneDrive - Marquette University/SCHEMA/Fig-enrichment_of_top_genes.png")
topmirage_enrichment
#dev.off()
```





```{r,  message=F, warning=F}
percentage_plot=data.frame(prop=c(percentage[2,], percentage[1,], percentage[4,]), genes=rep(c("Autism genes", "DD/ID genes", "SCZ GWAS genes"),3), group=rep(c("all genes", "DAWN genes", "DAWN new genes"), each=3))
topmirage_enrichment=
  ggplot(percentage_plot, aes(genes, prop)) + 
   geom_bar(aes(fill = group), stat = "identity", position = "dodge")+
   theme_classic()+
   xlab("")+ylab("Proportion of genes in the gene set")+
  theme(axis.text.y = element_text(size = 13))+ # this is for numbers on y axis 
   ylim(c(0, 0.35))+
   theme(plot.title = element_text(hjust = 0.5, size=15, face="bold"))+
   ggtitle("Enrichment of DAWN genes with other gene sets")+
    theme(legend.title = element_blank())+
   geom_signif(annotations ="0.113", y_position = 0.175  ,xmin=0.7, xmax=1.3, size=1, vjust=-0.5, textsize=5.0)+ # textsize is for text over the bar; size is for width of lines 
  geom_signif(annotations ="0.004", y_position = 0.135  ,xmin=0.7, xmax=1, size=1, vjust=-0.5, textsize=5.0)+ # textsize is for text over the bar; size is for width of lines
     geom_signif(annotations ="8e-05", y_position = 0.225  ,xmin=1.7, xmax=2, size=1, vjust=-0.5, textsize=5.0)+
    geom_signif(annotations ="0.541", y_position = 0.26  ,xmin=1.7, xmax=2.3, size=1, vjust=-0.5, textsize=5.0)+
   geom_signif(annotations ="0.033", y_position = 0.05  ,xmin=2.7, xmax=3, size=1, vjust=-0.5, textsize=5.0)+
     geom_signif(annotations ="1.000", y_position = 0.09  ,xmin=2.7, xmax=3.3, size=1, vjust=-0.5, textsize=5.0)+
   theme(legend.position="bottom")+
   theme(axis.text.x = element_text(angle = 0, vjust = 0.5, size=15))+
 theme(text = element_text(size=15))+ # size is for labels on x axis 
  theme(axis.title = element_text(size = 15))+ # Font Size of Axis Titles
 theme(legend.spacing.x = unit(0.5, 'cm')) # add space between symbols in legend 

#png("C:/Users/hans/OneDrive - Marquette University/SCHEMA/Fig-enrichment_of_top_genes.png")
topmirage_enrichment
#dev.off()
###########  p values are pvalue_dawn; pvalue_dawn_new


```



## overlap with MIRAGE FDR genes 


```{r,message=F, warning=F, eval=T}

##############
##########
rm("BF.PP.gene")
#BF.PP.gene=as_tibble(read.csv("C:\\han\\Dataset\\SCHEMA\\MIRAGE_on_meta_sample_estimate_delta_3vargroups_20231027.csv", header=T))
load("C:\\han\\Dataset\\SCHEMA\\MIRAGE_on_meta_sample_estimate_delta_2vargroups_for_constraint_gene_20231120.RData")

fdr_function=function(fdr_threshold, data)
{
tau=seq(0, 0.999, by=0.001)
num_pred=NULL
false_disc=NULL
FDR=NULL
for (i in 1:length(tau))
{
num_pred[i]=sum(ifelse(data$post_prob>tau[i], 1, 0))
false_disc[i]=sum((1-data$post_prob)*ifelse(data$post_prob>tau[i], 1, 0))
FDR[i]=false_disc[i]/num_pred[i]
}
tau_fdr=tibble(tau=tau, fdr=FDR)%>%drop_na()  # drop rows with NA
tau_at_threshold=tau_fdr%>%filter(fdr<=fdr_threshold)
if (nrow(tau_at_threshold)==0)
  return("No gene is selected")
if (nrow(tau_at_threshold)>0)
{
  tau_threshold=tau_at_threshold$tau[1]
  return(data%>% filter(post_prob>tau_threshold))
}

}
```

```{r,message=F, warning=F, eval=T}
input_data=BF.PP.gene
colnames(input_data)[3]="post_prob"
FDR_gene=fdr_function(fdr_threshold = 0.05, input_data)
```


```{r, message=F, warning=F}
intersect(FDR_gene$Gene, genes_pfc35)  # overlapping genes between 110 MIRAGE FDR genes and 46 DAWN genes
setdiff(intersect(FDR_gene$Gene, genes_pfc35), SCHEMA_FDR_gene)  # 18 new DAWN genes overlapping with 110 MIRAGE FDR genes 
```
