==> analysis/part6.Rmd <==
---
title: "functional enrichment analysis"
output: html_document
date: "2024-06-13"
---


## GO enrichment 


```{r, echo=F, message=F, warning=F}
#https://biostatsquid.com/volcano-plots-r-tutorial/
library(ggrepel) # for nice annotations

go_enrichment=data.frame(terms=c("GO:0086019
", "GO:0086065", "GO:0034702", "GO:1902495", "GO:1990351", "GO:0022839", "GO:0022836", "GO:0022890", "GO:0015318"), fold_change=c(63.56, 39.91, 18.26, 13.69, 13.41, 10.09, 10.09, 8.09, 6.61), pvalue=c(0.00401, 0.0274, 0.000389, 0.00207, 0.00234, 0.015, 0.015, 0.013, 0.045))

ggplot(data = go_enrichment, aes(x = log2(fold_change), y = -log10(pvalue), label = terms)) +
  geom_hline(yintercept = -log10(0.05), col = "red", linetype = 'dashed') +
         geom_point(aes(size = log2(fold_change)))+
  xlim(c(0,7))+
         labs(color = 'Severe', #legend_title, 
       x = expression("log"[2]*" (fold change)"), y = expression("-log"[10]*"(p value)")) +
            theme(
              axis.title.y = element_text(face = "bold",  size =14, color = 'black'),
              axis.text.y =element_text(size=12), 
              axis.title.x = element_text(hjust = 0.5, face = "bold",  size = 14, color = 'black'),
              axis.text.x =element_text(size=12),
              plot.title = element_text(hjust = 0.5)
            )+
    geom_text_repel(max.overlaps = Inf)
```




## synGO enrichment


```{r, echo=F, message=F, warning=F, eval=F}
load("../data/dat_pfc35_20240328.RData")
write.table(colnames(dat_pfc35), file="C:\\Users\\hans\\OneDrive - Marquette University\\SCZ-DAWN\\3100Genes.txt",row.names=F, col.names = F, quote = F) # output this gene list as input for GO enrichment analysis
```
